Moved rebuild command into utils/
